fix issue with accidentally still casting the mouse world coordinates
to an int, which had..obvious repercussions.

+++ src/player.h
@@ -52,8 +52,8 @@ public:
     /**
      * Server side only, to asociate and keep track of each player's (clients) mouse position
      */
-    void setMousePositionWorldCoords(int32_t x, int32_t y) {
-        m_mousePositionWorldCoords = glm::ivec2(x, y);
+    void setMousePositionWorldCoords(float x, float y) {
+        m_mousePositionWorldCoords = glm::vec2(x, y);
     }
     glm::vec2 mousePositionWorldCoords() {
         return m_mousePositionWorldCoords;
